code/requirements.txt
alembic==1.2.0
art==3.7
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
configparser==4.0.2
coverage==4.5.3
cx-Oracle==7.0.0
cycler==0.10.0
Cython==0.29.3
dask==1.1.3
databricks-cli==0.9.0
decorator==4.3.2
docker==4.0.2
entrypoints==0.3
Flask==1.0.2
gitdb2==2.0.5
GitPython==3.0.2
gorilla==0.3.0
gunicorn==19.9.0
idna==2.8
imgaug==0.2.6
itsdangerous==1.1.0
kiwisolver==1.0.1
llvmlite==0.29.0
lxml==4.3.3
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.0.3
mlflow==1.2.0
networkx==2.2
numba==0.45.1
numpy==1.16.0
opencv-python==4.1.1.26
openslide-python==1.1.1
pandas==0.24.1
Pillow==5.4.1
protobuf==3.9.1
pyclipper==1.1.0.post1
pycm==2.3
pycurl==7.43.0
pyparsing==2.3.1
python-dateutil==2.7.5
python-editor==1.0.4
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1.2
pyzmq==17.1.2
querystring-parser==1.2.4
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
simplejson==3.16.0
six==1.12.0
sklearn==0.0
smmap2==2.0.5
SQLAlchemy==1.3.8
sqlparse==0.3.0
tabulate==0.8.3
tifffile==2019.1.300
toolz==0.9.0
torch==1.0.0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.29.1
urllib3==1.24.1
visdom==0.1.8.8
webcolors==1.8.1
websocket-client==0.54.0
Werkzeug==0.14.1
xlrd==1.2.0
xmltodict==0.12.0